updated names of Route and Path classes to RouteTemplate and Route
